+++ src/components/__tests__/CurrencyInput-fixedDecimalLength.spec.tsx
@@ -1,11 +1,11 @@
-import { shallow } from 'enzyme';
 import React from 'react';
+import '@testing-library/jest-dom';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import CurrencyInput from '../CurrencyInput';
 
-const id = 'validationCustom01';
-
-describe('<CurrencyInput /> component > fixedDecimalLength', () => {
-  const onBlurValueSpy = jest.fn();
+describe('<CurrencyInput/> fixedDecimalLength', () => {
+  const onValueChangeSpy = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -13,48 +13,45 @@ describe('<CurrencyInput /> component > fixedDecimalLength', () => {
 
   describe('fixedDecimalLength', () => {
     it('should convert value on blur if fixedDecimalLength specified', () => {
-      const view = shallow(
+      render(
         <CurrencyInput
-          id={id}
           prefix="$"
-          onBlurValue={onBlurValueSpy}
+          onValueChange={onValueChangeSpy}
           fixedDecimalLength={3}
           defaultValue={123}
         />
       );
 
-      const input = view.find(`#${id}`);
-      expect(input.prop('value')).toBe('$123');
+      expect(screen.getByRole('textbox')).toHaveValue('$123');
 
-      view.find(`#${id}`).simulate('blur', { target: { value: '$123' } });
-      expect(onBlurValueSpy).toBeCalledWith('1.230', undefined);
+      screen.getByRole('textbox').focus();
+      userEvent.tab();
 
-      const updatedView = view.update();
-      expect(updatedView.find(`#${id}`).prop('value')).toBe('$1.230');
+      expect(onValueChangeSpy).toBeCalledWith('1.230', undefined);
+
+      expect(screen.getByRole('textbox')).toHaveValue('$1.230');
     });
 
-    it('should work with precision and decimalSeparator', () => {
-      const view = shallow(
+    it('should work with decimalScale and decimalSeparator', () => {
+      render(
         <CurrencyInput
-          id={id}
           prefix="$"
-          onBlurValue={onBlurValueSpy}
+          onValueChange={onValueChangeSpy}
           fixedDecimalLength={3}
-          groupSeparator="."
           decimalSeparator=","
-          precision={2}
+          decimalScale={2}
           defaultValue={123}
         />
       );
 
-      const input = view.find(`#${id}`);
-      expect(input.prop('value')).toBe('$123');
+      expect(screen.getByRole('textbox')).toHaveValue('$123');
 
-      view.find(`#${id}`).simulate('blur', { target: { value: '$123' } });
-      expect(onBlurValueSpy).toBeCalledWith('1,23', undefined);
+      screen.getByRole('textbox').focus();
+      userEvent.tab();
 
-      const updatedView = view.update();
-      expect(updatedView.find(`#${id}`).prop('value')).toBe('$1,23');
+      expect(onValueChangeSpy).toBeCalledWith('1,23', undefined);
+
+      expect(screen.getByRole('textbox')).toHaveValue('$1,23');
     });
   });
 });