+++ src/components/utils/index.ts
@@ -1,5 +1,7 @@
 export * from './cleanValue';
 export * from './fixedDecimalValue';
 export * from './formatValue';
+export * from './getLocaleConfig';
+export * from './getSuffix';
 export * from './isNumber';
 export * from './padTrimValue';