+++ src/examples/Example1.tsx
@@ -7,16 +7,14 @@ export const Example1: FC = () => {
 
   const [errorMessage, setErrorMessage] = useState('');
   const [className, setClassName] = useState('');
-  const [value, setValue] = useState<string | number>(999.99);
+  const [value, setValue] = useState<string | number>(123.45);
   const [rawValue, setRawValue] = useState<string | undefined>(' ');
-  const [rawBlurValue, setRawBlurValue] = useState<string | undefined>(' ');
 
   /**
    * Handle validation
    */
-  const validateValue = (value: string | undefined): void => {
-    const rawValue = value === undefined ? 'undefined' : value;
-    setRawValue(rawValue || ' ');
+  const handleOnValueChange = (value: string | undefined): void => {
+    setRawValue(value === undefined ? 'undefined' : value || ' ');
 
     if (!value) {
       setClassName('');
@@ -41,11 +39,6 @@ export const Example1: FC = () => {
     setValue(value);
   };
 
-  const handleOnBlurValue = (value: string | undefined) => {
-    const rawBlurValue = value === undefined ? 'undefined' : value;
-    setRawBlurValue(rawBlurValue || ' ');
-  };
-
   return (
     <div className="row">
       <div className="col-12 mb-4">
@@ -55,7 +48,6 @@ export const Example1: FC = () => {
         <ul>
           <li>{`'£'`} prefix</li>
           <li>Allows decimals (up to 2 decimal places)</li>
-          <li>Has default value (999.99)</li>
           <li>Value is set programmatically (passed in via props)</li>
         </ul>
 
@@ -66,13 +58,11 @@ export const Example1: FC = () => {
               <CurrencyInput
                 id="validationCustom01"
                 name="input-1"
-                defaultValue={999.99}
                 className={`form-control ${className}`}
                 value={value}
-                onChange={validateValue}
-                onBlurValue={handleOnBlurValue}
+                onValueChange={handleOnValueChange}
+                placeholder="Please enter a number"
                 prefix={prefix}
-                precision={2}
                 step={1}
               />
               <div className="invalid-feedback">{errorMessage}</div>
@@ -81,12 +71,8 @@ export const Example1: FC = () => {
               <pre className="h-100 p-3 bg-dark text-white">
                 <div className="row">
                   <div className="col-6">
-                    <div className="text-muted mr-3">onChange:</div>
+                    <div className="text-muted mr-3">onValueChange:</div>
                     {rawValue}
-                  </div>
-                  <div className="col-6">
-                    <div className="text-muted mr-3">onBlurValue:</div>
-                    {rawBlurValue}
                   </div>
                 </div>
               </pre>